Feature/backend/library api

See merge request s08-webmobile2-sub2/S08P12A209!229

+++ BE/src/main/java/com/seoro/seoro/service/Library/LibraryServiceImpl.java
@@ -1,23 +1,12 @@
 package com.seoro.seoro.service.Library;
 
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.URL;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
-import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 
-import com.seoro.seoro.domain.dto.Book.BookDetailDto;
+import org.springframework.stereotype.Service;
+
 import com.seoro.seoro.domain.dto.Book.BookReportDto;
 import com.seoro.seoro.domain.dto.Book.OwnBookDto;
 import com.seoro.seoro.domain.dto.Book.OwnCommentDto;
@@ -115,8 +104,8 @@ public class LibraryServiceImpl implements LibraryService {
 		boolean isFollowing;
 		if(!isOwn) {
 			// 팔로잉에 유저 아이디 팔로워에 내 아이디면 팔로잉한 유저
-			Optional<Friend> friend = friendRepository.findByFollowerAndFollowing(me, memberId);
-			if(friend.isPresent()) {
+			Friend friend = friendRepository.findByFollowerAndFollowing(me, memberId).orElse(null);
+			if(friend != null) {
 				isFollowing = true;
 			} else {
 				isFollowing = false;
@@ -214,7 +203,6 @@ public class LibraryServiceImpl implements LibraryService {
 		return responseDto;
 	}
 
-	// 수정
 	@Override
 	public ResultResponseDto removeReadBook(Long memberId, String isbn) {
 		log.info("memberId: " + memberId);
@@ -327,16 +315,18 @@ public class LibraryServiceImpl implements LibraryService {
 
 	@Override
 	public BookReportDto viewBookReport(Long bookReportId) {
-		// 안 쓰는 코드
-		// Optional<BookReport> bookReport = bookReportRepository.findById(bookReportId);
-		// BookReport responseBookReport = bookReport.get();
-		//
-		// BookReportDto responsetDto = new BookReportDto
-		// 	(responseBookReport.getReadBook().getReadBookId(), responseBookReport.getBookReportTitle(), responseBookReport.getBookReportContent());
-		return null;
+		BookReport bookReport = bookReportRepository.findById(bookReportId).orElse(null);
+
+		if(bookReport == null) {
+			BookReportDto responsetDto = new BookReportDto();
+			responsetDto.setResult(false);
+			return responsetDto;
+		}
+
+		BookReportDto responsetDto = new BookReportDto(bookReport);
+		return responsetDto;
 	}
 
-	// 수정
 	@Override
 	public ResultResponseDto modifyBookReport(BookReportDto requestDto) {
 		ResultResponseDto responseDto = new ResultResponseDto();
@@ -354,8 +344,9 @@ public class LibraryServiceImpl implements LibraryService {
 		// 이미지 수정 추가
 
 		BookReport newBookReport = BookReport.builder()
-			.bookReportId(bookReport.getBookReportId())
+			.bookReportId(requestDto.getBookReportId())
 			.readBook(readBook)
+			.member(bookReport.getMember())
 			.bookReportTitle(requestDto.getBookReportTitle())
 			.bookReportContent(requestDto.getBookReportContent())
 			.build();
@@ -366,7 +357,6 @@ public class LibraryServiceImpl implements LibraryService {
 		return responseDto;
 	}
 
-	// 수정
 	@Override
 	public ResultResponseDto removeBookReport(Long bookReportId) {
 		ResultResponseDto responseDto = new ResultResponseDto();
@@ -383,32 +373,56 @@ public class LibraryServiceImpl implements LibraryService {
 		return responseDto;
 	}
 
-	// 수정
 	@Override
-	public LibraryDto makeFriend(Long memberId, Long meId) {
-		LibraryDto responseDto = libraryMain(memberId, meId);
-		Member member = memberRepository.findByMemberId(meId);
+	public ResultResponseDto makeFriend(Long memberId, Long meId) {
+		ResultResponseDto responseDto = new ResultResponseDto();
+
+		Member member = memberRepository.findByMemberId(memberId);
+		Member me = memberRepository.findByMemberId(meId);
+		if(member == null || member == null) {
+			responseDto.setResult(false);
+			responseDto.setMessege("찾는 회원이 없습니다.");
+			return responseDto;
+		}
+
+		Friend checkFriend = friendRepository.findByFollowerAndFollowing(me, memberId).orElse(null);
+		if(checkFriend != null) {
+			responseDto.setResult(false);
+			responseDto.setMessege("이미 추가한 친구입니다.");
+			return responseDto;
+		}
 
 		Friend friend = Friend.builder()
-			.follower(member)
+			.follower(me)
 			.following(memberId)
 			.build();
 		friendRepository.save(friend);
-		responseDto.setMyFollowings(responseDto.getMyFollowers() + 1);
+		responseDto.setResult(true);
 
 		return responseDto;
 	}
 
-	// 수정
 	@Override
-	public LibraryDto removeFriend(Long memberId, Long meId) {
-		LibraryDto responseDto = libraryMain(memberId, meId);
-		Member member = memberRepository.findByMemberId(meId);
+	public ResultResponseDto removeFriend(Long memberId, Long meId) {
+		ResultResponseDto responseDto = new ResultResponseDto();
 
-		Friend friend = friendRepository.findByFollowerAndFollowing(member, memberId).orElseThrow(() -> new NoSuchElementException("친구가 아닙니다"));
+		Member member = memberRepository.findByMemberId(memberId);
+		Member me = memberRepository.findByMemberId(meId);
+		if(member == null || member == null) {
+			responseDto.setResult(false);
+			responseDto.setMessege("찾는 회원이 없습니다.");
+			return responseDto;
+		}
+
+		Friend friend = friendRepository.findByFollowerAndFollowing(me, memberId).orElse(null);
+		if(friend == null) {
+			responseDto.setResult(false);
+			responseDto.setMessege("친구가 없습니다.");
+			return responseDto;
+		}
 
 		friendRepository.delete(friend);
-		responseDto.setMyFollowings(responseDto.getMyFollowers() - 1);
+		responseDto.setResult(true);
 
 		return responseDto;
 	}
